+++ libs/docs/platform/table/e2e/table-p13-dialog.e2e-spec.ts
@@ -1,6 +1,7 @@
 import {
     browserIsSafari,
     click,
+    doesItExist,
     getElementArrayLength,
     getText,
     refreshPage,
@@ -47,9 +48,12 @@ describe('Table component test suite', () => {
         dropdownOption,
         dialogButton,
         allInputFields,
-        ellipsisButton,
+        toolbarButton,
         expandedOption,
-        buttonFilter
+        buttonFilter,
+        slimArrowRight,
+        dialogDecisiveButton,
+        messageBoxHeader
     } = tablePage;
 
     beforeAll(async () => {
@@ -78,7 +82,7 @@ describe('Table component test suite', () => {
         });
 
         it('should check sorting of columns', async () => {
-            await tablePage.checkSortingColumns(tableP13ColumnsExample, ellipsisButton);
+            await tablePage.checkSortingColumns(tableP13ColumnsExample, toolbarButton);
         });
     });
 
@@ -89,7 +93,7 @@ describe('Table component test suite', () => {
 
         it('should check sorting ascending and descending by name', async () => {
             await scrollIntoView(tableP13SortExample);
-            await click(tableP13SortExample + ellipsisButton);
+            await click(tableP13SortExample + toolbarButton);
             await click(popoverDropdownButton);
             await click(buttonSortedBy);
             await click(footerButtonOk);
@@ -104,7 +108,7 @@ describe('Table component test suite', () => {
 
         it('should check sorting ascending and descending by price', async () => {
             await scrollIntoView(tableP13SortExample);
-            await click(tableP13SortExample + ellipsisButton);
+            await click(tableP13SortExample + toolbarButton);
             await click(buttonAdd);
             await click(buttonRemove);
             await click(popoverDropdownButton);
@@ -125,7 +129,7 @@ describe('Table component test suite', () => {
         });
 
         it('should check sorting of columns', async () => {
-            await tablePage.checkSortingColumns(tableP13SortExample, ellipsisButton, 1);
+            await tablePage.checkSortingColumns(tableP13SortExample, toolbarButton, 1);
         });
     });
 
@@ -136,7 +140,7 @@ describe('Table component test suite', () => {
 
         it('should check filtering with include and exclude', async () => {
             await scrollIntoView(tableP13FilterExample);
-            await click(tableP13FilterExample + ellipsisButton);
+            await click(tableP13FilterExample + toolbarButton);
             await setValue(dialogInput, astroTestText);
             await click(expandedButton, 1);
             await click(popoverDropdownButton, 2);
@@ -156,7 +160,16 @@ describe('Table component test suite', () => {
         });
 
         it('should check sorting of columns', async () => {
-            await tablePage.checkSortingColumns(tableP13FilterExample, ellipsisButton, 1);
+            await tablePage.checkSortingColumns(tableP13FilterExample, toolbarButton, 1);
+        });
+        it('should throw warning when the user attempts to include/exclude the same rule', async () => {
+            await scrollIntoView(tableP13FilterExample);
+            await click(tableP13FilterExample + toolbarButton);
+            await setValue(dialogInput, 'x');
+            await click(slimArrowRight);
+            await setValue(dialogInput, 'x', 1);
+            await click(dialogDecisiveButton);
+            await expect(await doesItExist(messageBoxHeader)).toBe(true);
         });
     });
 
@@ -171,7 +184,7 @@ describe('Table component test suite', () => {
         });
 
         it('should check sorting of columns', async () => {
-            await tablePage.checkSortingColumns(tableP13GroupExample, ellipsisButton, 1);
+            await tablePage.checkSortingColumns(tableP13GroupExample, toolbarButton, 1);
         });
     });