+++ CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.49.1-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.49.0...v0.49.1-rc.0) (2024-03-21)
+
+
+### Bug Fixes
+
+* **platform:** vhd keyboard improvements ([#11562](https://github.com/SAP/fundamental-ngx/issues/11562)) ([09a73e7](https://github.com/SAP/fundamental-ngx/commit/09a73e74f965fa0a3d1cde80a90f1cd5b3b83eb2))
+
+
+
+
+
+# [0.49.0](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.68...v0.49.0) (2024-03-21)
+
+**Note:** Version bump only for package fundamental-ngx
+
+
+
+
+
+# [0.49.0-rc.68](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.67...v0.49.0-rc.68) (2024-03-21)
+
+
+### Bug Fixes
+
+* **platform:** add validator input for p13 filter component ([#11520](https://github.com/SAP/fundamental-ngx/issues/11520)) ([a3b23ff](https://github.com/SAP/fundamental-ngx/commit/a3b23ffd495c4fb1afbc75938f5988d457406e7c))
+
+
+
+
+
 # [0.49.0-rc.67](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.66...v0.49.0-rc.67) (2024-03-21)